+++ src/commonMain/kotlin/com/insanusmokrassar/sdi/Module.kt
@@ -1,5 +1,0 @@
-package com.insanusmokrassar.sdi
-
-import kotlinx.serialization.ContextualSerialization
-
-class Module internal constructor(base: Map<String, @ContextualSerialization Any>) : Map<String, Any> by base