+++ src/jvmMain/kotlin/com/insanusmokrassar/sdi/utils/resolveKClassByPackageName.kt
@@ -1,5 +1,0 @@
-package com.insanusmokrassar.sdi.utils
-
-import kotlin.reflect.KClass
-
-actual fun resolveKClassByPackageName(packageName: String): KClass<*> = Class.forName(packageName).kotlin